refactor: centralize error suggestion handling
